DRY up audit components, fix path on Docs, fix database for testing/pipeline
